+++ packages/emotion/src/index.js
@@ -6,8 +6,20 @@ export const sheet = new StyleSheet()
 sheet.inject()
 const stylisOptions = { keyframe: false }
 
-const stylis = new Stylis(stylisOptions)
-const keyframeStylis = new Stylis(stylisOptions)
+if (process.env.NODE_ENV !== 'production') {
+  stylisOptions.compress = false
+}
+
+let stylis = new Stylis(stylisOptions)
+
+const externalStylisPlugins = []
+
+const use = stylis.use
+
+export const useStylisPlugin = plugin => {
+  externalStylisPlugins.push(plugin)
+  use(null)(externalStylisPlugins)(insertionPlugin)
+}
 
 export let registered = {}
 
@@ -20,30 +32,76 @@ export function flush() {
   sheet.inject()
 }
 
-function insertionPlugin(context, content, selector, parent) {
-  switch (context) {
-    case 2: {
-      if (parent[0] === selector[0]) {
-        break
-      }
-    }
-    // after an at rule block
-    case 3: // eslint-disable-line no-fallthrough
-      sheet.insert(`${selector.join(',')}{${content}}`)
-  }
+let currentSourceMap = ''
+let queue = []
+
+function insertRule(rule) {
+  sheet.insert(rule, currentSourceMap)
 }
 
-function keyframeInsertionPlugin(context, content, selector) {
-  if (context === 3) {
-    sheet.insert(
-      `${selector[0].replace('keyframes', '-webkit-keyframes')}{${content}}`
-    )
-    sheet.insert(`${selector[0]}{${content}}`)
+function insertionPlugin(
+  context,
+  content,
+  selectors,
+  parent,
+  line,
+  column,
+  length,
+  id
+) {
+  switch (context) {
+    case -2: {
+      queue.forEach(insertRule)
+      queue = []
+      break
+    }
+
+    case 2: {
+      if (id === 0) {
+        const joinedSelectors = selectors.join(',')
+        const rule = `${joinedSelectors}{${content}}`
+        if (parent.join(',') === joinedSelectors || parent[0] === '') {
+          queue.push(rule)
+        } else {
+          queue.unshift(rule)
+        }
+      }
+      break
+    }
+    // after an at rule block
+    case 3: {
+      let chars = selectors.join('')
+      const second = chars.charCodeAt(1)
+      let child = content
+      switch (second) {
+        // s upports
+        case 115:
+        // d ocument
+        // eslint-disable-next-line no-fallthrough
+        case 100:
+        // m edia
+        // eslint-disable-next-line no-fallthrough
+        case 109: {
+          queue.push(chars + '{' + child + '}')
+          break
+        }
+        // k eyframes
+        case 107: {
+          chars = chars.substring(1)
+          child = chars + '{' + child + '}'
+          queue.push('@-webkit-' + child)
+          queue.push('@' + child)
+          break
+        }
+        default: {
+          queue.push(chars + child)
+        }
+      }
+    }
   }
 }
 
 stylis.use(insertionPlugin)
-keyframeStylis.use(keyframeInsertionPlugin)
 
 function flatten(inArr) {
   let arr = []
@@ -59,15 +117,22 @@ function handleInterpolation(
   interpolation: any,
   couldBeSelectorInterpolation: boolean
 ) {
-  if (typeof interpolation === 'object') {
-    return createStringFromObject(interpolation)
-  }
   if (
     interpolation === undefined ||
     interpolation === null ||
-    interpolation === false
-  )
+    typeof value === 'boolean'
+  ) {
     return ''
+  }
+
+  if (typeof interpolation === 'function') {
+    return handleInterpolation(interpolation())
+  }
+
+  if (typeof interpolation === 'object') {
+    return createStringFromObject(interpolation)
+  }
+
   if (
     couldBeSelectorInterpolation === false &&
     registered[interpolation] !== undefined
@@ -93,7 +158,12 @@ const processStyleValue = (key, value) => {
   return value
 }
 
+const objectToStringCache = new WeakMap()
+
 function createStringFromObject(obj) {
+  if (objectToStringCache.has(obj)) {
+    return objectToStringCache.get(obj)
+  }
   let string = ''
 
   if (Array.isArray(obj)) {
@@ -112,10 +182,12 @@ function createStringFromObject(obj) {
           )};`
         }
       } else {
-        string += `${key}{${createStringFromObject(obj[key])}}`
+        string += `${key}{${handleInterpolation(obj[key])}}`
       }
     })
   }
+  objectToStringCache.set(obj, string)
+
   return string
 }
 
@@ -126,10 +198,7 @@ function isLastCharDot(string) {
 function createStyles(strings, ...interpolations) {
   let stringMode = true
   let styles = ''
-  if (
-    (strings !== undefined && strings.raw === undefined) ||
-    strings === undefined
-  ) {
+  if (strings == null || strings === undefined || strings.raw === undefined) {
     stringMode = false
     styles = handleInterpolation(strings, false)
   } else {
@@ -141,7 +210,19 @@ function createStyles(strings, ...interpolations) {
       styles += strings[i + 1]
     }
   })
+
   return styles
+}
+
+if (process.env.NODE_ENV !== 'production') {
+  const sourceMapRegEx = /\/\*#\ssourceMappingURL=data:application\/json;\S+\s+\*\/\s+\/\*@\ssourceURL=\S+\s+\*\//
+  const oldStylis = stylis
+  stylis = (selector, styles) => {
+    const result = sourceMapRegEx.exec(styles)
+    currentSourceMap = result ? result[0] : ''
+    oldStylis(selector, styles)
+    currentSourceMap = ''
+  }
 }
 
 export function css(...args) {
@@ -167,18 +248,12 @@ export function injectGlobal(...args) {
   }
 }
 
-export function hydrate(ids) {
-  ids.forEach(id => {
-    inserted[id] = true
-  })
-}
-
 export function keyframes(...args) {
   const styles = createStyles(...args)
   const hash = hashString(styles)
   const name = `animation-${hash}`
   if (inserted[hash] === undefined) {
-    keyframeStylis('', `@keyframes ${name}{${styles}}`)
+    stylis('', `@keyframes ${name}{${styles}}`)
     inserted[hash] = true
   }
   return name
@@ -188,7 +263,7 @@ export function fontFace(...args) {
   const styles = createStyles(...args)
   const hash = hashString(styles)
   if (inserted[hash] === undefined) {
-    sheet.insert(`@font-face{${styles}}`)
+    stylis('', `@font-face{${styles}}`)
     inserted[hash] = true
   }
 }
@@ -206,7 +281,7 @@ export function getRegisteredStyles(registeredStyles, classNames) {
   return rawClassName
 }
 
-export function merge(className) {
+export function merge(className, sourceMap) {
   const registeredStyles = []
 
   const rawClassName = getRegisteredStyles(registeredStyles, className)
@@ -214,5 +289,11 @@ export function merge(className) {
   if (registeredStyles.length < 2) {
     return className
   }
-  return rawClassName + css(...registeredStyles)
+  return rawClassName + css(registeredStyles, sourceMap)
+}
+
+export function hydrate(ids) {
+  ids.forEach(id => {
+    inserted[id] = true
+  })
 }