+++ packages/emotion/test/css.test.js
@@ -1,6 +1,13 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { css, flush } from 'emotion'
+import { css, flush, sheet, useStylisPlugin } from 'emotion'
+import { transform } from 'cssjanus'
+
+useStylisPlugin(function(context, content) {
+  if (context === 2) {
+    return transform(content)
+  }
+})
 
 describe('css', () => {
   test('float property', () => {
@@ -29,6 +36,25 @@ describe('css', () => {
     expect(tree).toMatchSnapshot()
   })
 
+  test('falsy value in nested selector on object', () => {
+    const cls1 = css({
+      ':hover': {
+        display: null,
+        color: 'hotpink'
+      }
+    })
+    const tree = renderer.create(<div className={cls1} />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+  test('boolean as value', () => {
+    const cls1 = css({
+      display: 'flex',
+      color: false,
+      backgroundColor: true
+    })
+    const tree = renderer.create(<div className={cls1} />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
   test('auto px', () => {
     const cls1 = css({ display: 'flex', flex: 1, fontSize: 10 })
     const tree = renderer.create(<div className={cls1} />).toJSON()
@@ -85,6 +111,18 @@ describe('css', () => {
       height: 50,
       width: 20
     })
+    const tree = renderer.create(<div className={cls1} />).toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+
+  test('handles array of objects', () => {
+    const cls1 = css([
+      {
+        height: 50,
+        width: 20
+      },
+      null
+    ])
     const tree = renderer.create(<div className={cls1} />).toJSON()
     expect(tree).toMatchSnapshot()
   })
@@ -154,7 +192,7 @@ describe('css', () => {
     const tree = renderer.create(<div className={cls1} />).toJSON()
     expect(tree).toMatchSnapshot()
   })
-  test.skip('nested', () => {
+  test('nested', () => {
     const cls1 = css`
       color: yellow;
       & .some-class {
@@ -165,9 +203,6 @@ describe('css', () => {
         @media (max-width: 600px) {
           background-color: pink;
         }
-      }
-      &.another-class {
-        display: flex;
       }
     `
     const tree = renderer
@@ -180,6 +215,20 @@ describe('css', () => {
       )
       .toJSON()
     expect(tree).toMatchSnapshot()
+  })
+  test('explicit &', () => {
+    flush()
+    const cls1 = css`
+      &.another-class {
+        display: flex;
+      }
+    `
+    const tree = renderer
+      .create(<div className={`${cls1} another-class`} />)
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+    expect(sheet).toMatchSnapshot()
+    flush()
   })
   test('falsy property value in object', () => {
     const cls = css({ display: 'flex', backgroundColor: undefined })
@@ -232,6 +281,14 @@ describe('css', () => {
     const tree = renderer.create(<div className={cls1} />).toJSON()
     expect(tree).toMatchSnapshot()
   })
+
+  test('null value', () => {
+    const cls1 = css(null)
+    const cls2 = css`${() => null};`
+    expect(renderer.create(<div className={cls1} />).toJSON()).toMatchSnapshot()
+    expect(renderer.create(<div className={cls2} />).toJSON()).toMatchSnapshot()
+  })
+
   test('flushes correctly', () => {
     const cls1 = css`display: flex;`
     const tree = renderer.create(<div className={cls1} />).toJSON()
@@ -239,5 +296,39 @@ describe('css', () => {
     flush()
     const tree2 = renderer.create(<div className={cls1} />).toJSON()
     expect(tree2).toMatchSnapshot()
+  })
+  test('media query specificity', () => {
+    flush()
+    const cls = css`
+      width: 32px;
+      height: 32px;
+      border-radius: 50%;
+
+      @media (min-width: 420px) {
+        width: 96px;
+        height: 96px;
+      }
+    `
+
+    const tree = renderer.create(<div className={cls} />).toJSON()
+    expect(sheet).toMatchSnapshot()
+    expect(tree).toMatchSnapshot()
+    flush()
+  })
+  test('weakmap', () => {
+    const styles = { display: 'flex' }
+    const cls1 = css(styles)
+    const cls2 = css(styles)
+    const tree1 = renderer.create(<div className={cls1} />).toJSON()
+    expect(tree1).toMatchSnapshot()
+    const tree2 = renderer.create(<div className={cls2} />).toJSON()
+    expect(tree2).toMatchSnapshot()
+  })
+
+  test('return function in interpolation', () => {
+    const cls1 = css`color: ${() => 'blue'};`
+
+    const tree = renderer.create(<div className={cls1} />).toJSON()
+    expect(tree).toMatchSnapshot()
   })
 })